todo_project/requirements.txt
Django==5.0.4
psycopg2==2.9.9
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.31.0
djangorestframework-simplejwt==5.3.1